+++ source/Managers/MenuManager.h
@@ -39,8 +39,8 @@ class OnClickListenerAdapterList : public ListBox::ActionListener {
 public:
 	OnClickListenerAdapterList(ConfigScreen* screen) : config(screen) {
 	}
-	virtual void onClick(ListBox* listbox, UINT selectedIndex) override;
-	virtual void onHover(ListBox* listbox, short hoveredItemIndex) override;
+	virtual void onClick(ListBox* listbox, size_t selectedIndex) override;
+	virtual void onHover(ListBox* listbox, int hoveredItemIndex) override;
 private:
 	ConfigScreen* config;
 
@@ -50,8 +50,8 @@ class OnClickListenerDisplayModeList : public ComboBox::ActionListener {
 public:
 	OnClickListenerDisplayModeList(ConfigScreen* screen) : config(screen) {
 	}
-	virtual void onClick(ComboBox* combobox, UINT selectedIndex) override;
-	virtual void onHover(ComboBox * listbox, short hoveredItemIndex) override;
+	virtual void onClick(ComboBox* combobox, size_t selectedIndex) override;
+	virtual void onHover(ComboBox* listbox, int hoveredItemIndex) override;
 private:
 	ConfigScreen* config;
